+++ native/connector-rs/src/command.rs
@@ -1,17 +1,13 @@
 #![allow(non_snake_case)]
 
 use std::env;
-use std::process::Command;
 use std::str;
 
 extern crate json;
 use json::{object, JsonValue};
 
-extern crate log;
-use log::debug;
-
-use super::helpers::JsonGetters;
-use super::io;
+use crate::helpers::io;
+use crate::helpers::pipewire;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
@@ -23,7 +19,7 @@ fn GetVersion(payload: io::Payload) -> Result<JsonValue, String> {
 
 fn GetSessionType(payload: io::Payload) -> Result<JsonValue, String> {
   let session_type = match env::var_os("WAYLAND_DISPLAY") {
-    Some(val) => "wayland",
+    Some(_) => "wayland",
     None => "x11"
   };
 
@@ -33,29 +29,8 @@ fn GetSessionType(payload: io::Payload) -> Result<JsonValue, String> {
 }
 
 fn GetNodes(payload: io::Payload) -> Result<JsonValue, String> {
-  let dump_buffer = Command::new("pw-dump")
-    .output()
-    .expect("failed to execute process")
-    .stdout;
-
-  let dump_string = str::from_utf8(&dump_buffer).unwrap();
-  let dump = json::parse(dump_string).unwrap();
-
-  let dump_filtered = dump.members().filter(|&node| {
-    let media_class_result = node.get_fields_chain(vec!["info","props","media.class"]);
-    if let Err(e) = media_class_result {
-      debug! ("Error: {}", e);
-      return false;
-    }
-
-    let media_class = media_class_result.unwrap();
-    let media_class_string = media_class.as_str().unwrap();
-    media_class_string == "Stream/Output/Audio"
-  }).collect::<Vec<_>>();
-
-  let dump_converted = dump_filtered.iter().map(|&node| object! { "properties": node["info"]["props"].to_owned() }).collect::<Vec<_>>();
-
-  Ok(JsonValue::from(dump_converted))
+  let nodes = pipewire::get_output_nodes();
+  Ok(JsonValue::from(nodes))
 }
 
 fn StartPipewireScreenAudio(payload: io::Payload) -> Result<JsonValue, String> {
