+++ native/connector-rs/src/helpers.rs
@@ -6,6 +6,9 @@ use json::JsonValue;
 extern crate log;
 use log::debug;
 
+pub mod io;
+pub mod pipewire;
+
 pub trait JsonGetters {
   fn get_field_or_fail(&self, field: &str) -> Result<JsonValue,String>;
   fn get_fields_chain(&self, fields: Vec<&str>) -> Result<JsonValue,String>;
