+++ native/connector-rs/src/helpers/io.rs
